% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple_model.R
\docType{data}
\name{apple_model}
\alias{apple_model}
\title{Estimating closing stock price of Apple}
\format{
a \code{\link{lm}} object
}
\source{
Quandl
}
\usage{
apple_model
}
\description{
A simple Linear model that predicts the closing stock price of Apple from its opening price
}
